requirements.txt
piper-tts>=1.2.0
aiogram>=3.4.0
pyyaml>=6.0
httpx>=0.27.0
